verbs: Separate CM functions into their own file
