+++ Makefile.am
@@ -114,6 +114,7 @@ _verbs_files = \
 	prov/verbs/src/fi_verbs.h \
 	prov/verbs/src/fi_verbs.c \
 	prov/verbs/src/verbs_atomic.c \
+	prov/verbs/src/verbs_cm.c \
 	prov/verbs/src/verbs_cq.c \
 	prov/verbs/src/verbs_eq.c
